+++ profile_bluesky/startup/instrument/startup/__init__.py
@@ -1,7 +1,7 @@
 from .check_python import *
 from .check_bluesky import *
 
-from .framework import *
+from .initialize import *
 from .user_dir import *
 from .metadata import *
 from .callbacks import *
