+++ profile_bluesky/startup/instrument/startup/callbacks.py
@@ -9,15 +9,15 @@ __all__ = [
     'newSpecFile',
 ]
 
+from ..session_logs import logger
+logger.info(__file__)
+
 import apstools.filewriters
 import apstools.utils
 import datetime
 import os
 
-from ..startup.framework import RE, callback_db
-
-from ..session_logs import logger
-logger.info(__file__)
+from .initialize import RE, callback_db
 
 # write scans to SPEC data file
 specwriter = apstools.filewriters.SpecWriterCallback()
