+++ profile_bluesky/startup/instrument/startup/user_dir.py
@@ -5,12 +5,12 @@ support a .ipython/user directory for user files
 
 __all__ = []
 
+from ..session_logs import logger
+logger.info(__file__)
+
 import IPython.paths
 import os
 import sys
-
-from ..session_logs import logger
-logger.info(__file__)
 
 user_dir = os.path.join(
         IPython.paths.get_ipython_dir(),
